% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{ab1991}
\alias{ab1991}
\title{Arellano-Bond (1991) Panel Data for U.K. Companies}
\format{
A data frame with 1,031 rows and 10 variables:
\describe{
\item{year}{Year of observation}
\item{emp}{Employment (number of employees). Number of U.K. employees
(Datastream variable 216)}
\item{wage}{Real wage per employee. A measure of average annual
remuneration per employee was constructed by dividing U.K.
remuneration (Datastream variable 214) by the number of U.K. employees.
This was adjusted to
take into account changes in average weekly hours worked in manufacturing
industries (manual and non-manual employees, 18 years and over, male and
female, all occupations-source:
Department of Employment Gazette, various issues). A measure of real
average hourly
remuneration was then obtained by deflating using an implicit value-added
price deflator. These
implicit price deflators were calculated for each of our sub-sectors of
manufacturing industry,
using the current price and constant price GDP data published in various
Blue Books.}
\item{cap}{Gross capital stock. Denoting the historic cost book value of
gross fixed assets (Datastream variable 330) by HCKt, we
obtain an estimate of the inflation-adjusted (or replacement cost) value of
gross fixed assets
$RCK_t$ using the formula $RCK_t = HCK_t \\times P_t / P_\{t-A\}$ where $P$
is a price index for investment goods and
$A$ is an estimate of the average age of gross fixed assets. An implicit
price deflator for gross fixed
investment by manufacturing industry was calculated using the current price
and constant price
gross fixed investment data published in Economic Trends Annual Supplement
(1986, p. 56). For
the purpose of this exercise a value of $A$ of 6 years was assumed. Our
estimates of the gross
capital stock at replacement cost are then expressed in constant prices
using our investment
goods deflator.}
\item{indoutpt}{Industry output. An index of value-added output at constant
factor cost was constructed for each of our 9 sub-
sectors of manufacturing industry, using data published in the Blue Book
(1986, Table 2.4). The
15 sub-sectors of manufacturing for which this data is reported were
combined into 9 using the weights given in the Blue Book.}
\item{n}{Log of employment}
\item{w}{Log of real wage}
\item{k}{Log of gross capital stock}
\item{ys}{Log of industry output}
\item{id}{Firm identification code}
}
}
\source{
Arellano, Manual and Bond, Stephen. (1991). "Some Tests of Specification for Panel Data:
Monte Carlo Evidence and an Application to Employment Equations," \emph{Review of Economic Studies}, 277-297.
Data available via Stata command \verb{webuse abdata}.
}
\usage{
ab1991
}
\description{
Observations for 140 U.K. companies from 1976 to 1984.
}
\details{
The principal data source used is company accounts from Datastream International which
provide accounts records of employment and remuneration (i.e., wage bill) for all U.K. quoted
companies from 1976 onwards. We have used a sample of 140 companies with operations
mainly in the U.K., whose main activity is manufacturing and for which we have at least 7
continuous observations during the period 1976-1984. Where more than 7 observations are
available we have exploited this additional information, so that our sample has the unbalanced
structure described in Table A1.

As well as requiring at least 7 continuous observations, companies were excluded from our
sample for a number of reasons. We required complete records on a set of accounting variables
including gross fixed assets, investment, inventories and sales as well as employment and
remuneration. Companies that changed the date of their accounting year end by more than a
few days were excluded, so that our data all refer to 12 month periods. We also excluded
companies where either employment or one of our constructed measures of real wages, real
capital, real inventories or real sales jumped by more than a factor of 3 from one year to the
next. This filter will remove both those companies where data has been recorded erroneously
and those companies that have experienced major mergers. Finally, we restricted our attention
to companies that we could allocate to one of 9 broad sub-sectors of manufacturing industry
using Datastream's breakdown of total sales by product available from 1980 onwards.
}
\keyword{datasets}
