% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{ajr2001}
\alias{ajr2001}
\title{Acemoglou, Johnson, and Robinson (2001) Dataset on Colonial Origins of Comparative Development}
\format{
A data frame with 64 observations across 31 columns. Missing values are coded as NA.
\describe{
\item{longname}{Full country name}
\item{shortnam}{3-letter country name}
\item{logmort0}{Original log settler mortality}
\item{risk}{Expected risk}
\item{loggdp}{Log GDP per capita (PPP) in 1995}
\item{campaign}{Original mortality from campaign}
\item{source0}{Country source of own mortality rate}
\item{slave}{Data from forced labor}
\item{latitude}{Absolute value of latitude}
\item{neoeuro}{Indicator for 'New-Europes' (USA, CAN, AUS, NZL)}
\item{asia}{Indicator for Asia}
\item{africa}{Indicator for Africa}
\item{other}{Indicator for 'Other' continent (AUS, NZL, MLT)}
\item{edes1975}{Population of European descent, 1975}
\item{campaignsj}{Indicator for HKG and NZL}
\item{campaignsj2}{Extended recoding of campaign variable}
\item{mortnaval1}{Mortality using naval stations, method 1}
\item{logmortnaval1}{Log mortality using naval stations, method 1}
\item{mortnaval2}{Mortality using naval stations, method 2}
\item{logmortnaval2}{Log mortality using naval stations, method 2}
\item{mortjam}{Mortality benchmarked to the Caribbean}
\item{logmortjam}{Log mortality benchmarked to the Caribbean}
\item{logmortcap250}{Log mortality capped at 250}
\item{logmortjam250}{Log mortality benchmarked to the Caribbean, capped at 250}
\item{wandcafrica}{Indicator for West and Central Africa}
\item{malfal94}{Falciparum malaria index, 1994}
\item{wacacontested}{Indicator for contested observations in West and Central Africa}
\item{mortnaval2250}{Mortality using naval stations, capped at 250}
\item{logmortnaval2250}{Log mortality using naval stations, capped at 250}
\item{mortnaval1250}{Mortality using naval stations, capped at 250}
\item{logmortnaval1250}{Log mortality using naval stations, capped at 250}
}
}
\source{
Acemoglou, D., Johnson, S., & Robinson, J. A. (2001). "The Colonial Origins of Comparative Development: An Empirical Investigation." American Economic Review. Updated version (2012): "The Colonial Origins of Comparative Development: An Empirical Investigation: Reply." Data available at \url{https://www.aeaweb.org/aer/data/oct2012/20110390_data.zip}
}
\usage{
ajr2001
}
\description{
Colonial factors influencing economic development. Updated in 2012, it includes variables related to settler mortality, economic indicators, geographic data, and other socio-political metrics.
}
\keyword{datasets}
