% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{ak1991}
\alias{ak1991}
\title{Angrist and Krueger (1991) Dataset on Compulsory School Attendance}
\format{
A data frame with 329,509 observations across 10 columns.
\describe{
\item{ageq}{Age, including quarter of birth}
\item{edu}{Highest grade completed}
\item{logwage}{Log weekly earnings (in dollars)}
\item{married}{Married and with spouse present}
\item{qob}{Quarter of birth}
\item{black}{Black race}
\item{region}{Region of residence}
\item{smsa}{Works in a standard metropolitan statistical area (SMSA)}
\item{yob}{Year of birth}
\item{state}{State of birth}
}
}
\source{
Angrist, J., & Krueger, A. (1991). "Does Compulsory School Attendance
Affect Schooling and Earnings?" \emph{The Quarterly Journal of Economics}. Data
available at \url{http://economics.mit.edu/faculty/angrist/data1/data/angkru1991}
}
\usage{
ak1991
}
\description{
Schooling and earnings based on a subsample of men born 1930-1939 from the
1980 census extract. The dataset contains variables related to age,
education, earnings, marital status, and geographic information.
}
\keyword{datasets}
