% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{bmn2016}
\alias{bmn2016}
\title{Bernheim, Meer, and Novarro (2016) Dataset on Liquor Consumption and Commitment Opportunities}
\format{
A data frame with 1,722 observations and multiple variables related to liquor sales and policies.
\describe{
\item{id}{Unique identifier for each observation}
\item{state}{State name}
\item{year}{Year of observation}
\item{beeroffweekday}{Beer sales off-premises on weekdays}
\item{beerofffri}{Beer sales off-premises on Fridays}
\item{beeroffsat}{Beer sales off-premises on Saturdays}
\item{beeroffsun}{Beer sales off-premises on Sundays}
\item{wineoffweekday}{Wine sales off-premises on weekdays}
\item{wineofffri}{Wine sales off-premises on Fridays}
\item{wineoffsat}{Wine sales off-premises on Saturdays}
\item{wineoffsun}{Wine sales off-premises on Sundays}
\item{liqoffweekday}{Liquor sales off-premises on weekdays}
\item{liqofffri}{Liquor sales off-premises on Fridays}
\item{liqoffsat}{Liquor sales off-premises on Saturdays}
\item{liqoffsun}{Liquor sales off-premises on Sundays}
\item{beeronweekday}{Beer sales on-premises on weekdays}
\item{beeronfri}{Beer sales on-premises on Fridays}
\item{beeronsat}{Beer sales on-premises on Saturdays}
\item{beeronsun}{Beer sales on-premises on Sundays}
\item{wineonweekday}{Wine sales on-premises on weekdays}
\item{wineonfri}{Wine sales on-premises on Fridays}
\item{wineonsat}{Wine sales on-premises on Saturdays}
\item{wineonsun}{Wine sales on-premises on Sundays}
\item{liqonweekday}{Liquor sales on-premises on weekdays}
\item{liqonfri}{Liquor sales on-premises on Fridays}
\item{liqonsat}{Liquor sales on-premises on Saturdays}
\item{liqonsun}{Liquor sales on-premises on Sundays}
\item{beerethvol}{Beer ethanol volume}
\item{wineethvol}{Wine ethanol volume}
\item{liqethvol}{Liquor ethanol volume}
\item{beerage18}{Legal beer drinking age 18}
\item{beerage19}{Legal beer drinking age 19}
\item{beerage20}{Legal beer drinking age 20}
\item{beerage21}{Legal beer drinking age 21}
\item{wineage18}{Legal wine drinking age 18}
\item{wineage19}{Legal wine drinking age 19}
\item{wineage20}{Legal wine drinking age 20}
\item{wineage21}{Legal wine drinking age 21}
\item{liqage18}{Legal liquor drinking age 18}
\item{liqage19}{Legal liquor drinking age 19}
\item{liqage20}{Legal liquor drinking age 20}
\item{liqage21}{Legal liquor drinking age 21}
\item{unempw}{Unemployment rate}
\item{popover18}{Population over 18}
\item{beereth}{Beer ethanol content}
\item{wineeth}{Wine ethanol content}
\item{liqeth}{Liquor ethanol content}
\item{logliq}{Log of liquor sales}
\item{logwine}{Log of wine sales}
\item{logbeer}{Log of beer sales}
\item{expand}{Indicator for expansion of sales}
\item{restrict}{Indicator for restriction of sales}
\item{liqoffsunp}{Liquor off-premises sales on Sundays (percentage)}
\item{liqonsunp}{Liquor on-premises sales on Sundays (percentage)}
\item{beeroffinflows}{Beer off-premises inflows}
\item{wineoffinflows}{Wine off-premises inflows}
\item{liqoffinflows}{Liquor off-premises inflows}
\item{beeroninflows}{Beer on-premises inflows}
\item{wineoninflows}{Wine on-premises inflows}
\item{liqoninflows}{Liquor on-premises inflows}
\item{liqonoutflows}{Liquor on-premises outflows}
\item{liqoffoutflows}{Liquor off-premises outflows}
\item{beeronoutflows}{Beer on-premises outflows}
\item{beeroffoutflows}{Beer off-premises outflows}
\item{wineonoutflows}{Wine on-premises outflows}
\item{wineoffoutflows}{Wine off-premises outflows}
\item{liqonnotsun}{Liquor on-premises sales not on Sundays}
\item{liqoffnotsun}{Liquor off-premises sales not on Sundays}
\item{everonchange}{Indicator for any change in on-premises sales}
\item{everoffchange}{Indicator for any change in off-premises sales}
\item{winetax}{Wine tax}
\item{liquortax}{Liquor tax}
\item{beertax}{Beer tax}
\item{cigtax}{Cigarette tax}
\item{barsmokingban}{Ban on smoking in bars}
\item{restsmokingban}{Ban on smoking in restaurants}
}
}
\source{
Bernheim, B. D., Meer, J., & Novarro, N. K. (2016). "Do Consumers Exploit Commitment Opportunities? Evidence from Natural Experiments Involving Liquor Consumption." \emph{American Economic Journal: Economic Policy}, 8(1), 41-69. Data available at \url{https://www.aeaweb.org/articles?id=10.1257/pol.20130351}
}
\usage{
bmn2016
}
\description{
Liquor sale hours, alcohol consumption, and related state policies across 47 U.S. states from 1970 to 2007. The dataset includes off-premises and on-premises sales data for beer, wine, and liquor, along with other socioeconomic variables.
}
\keyword{datasets}
