% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{card1995}
\alias{card1995}
\title{Card (1995) Dataset on College Proximity and Return to Schooling}
\format{
A data frame with 3,613 observations and 52 variables. Missing values are coded as NA.
\describe{
\item{id}{Sequential ID, runs from 1 to 5225}
\item{nearc2}{Grew up near a 2-year college}
\item{nearc4}{Grew up near a 4-year college}
\item{nearc4a}{Grew up near a 4-year public college}
\item{nearc4b}{Grew up near a 4-year private college}
\item{ed76}{Education level in 1976}
\item{ed66}{Education level in 1966}
\item{age76}{Age in 1976}
\item{daded}{Father's education (missing values are set to average)}
\item{nodaded}{Indicator if father's education is imputed}
\item{momed}{Mother's education}
\item{nomomed}{Indicator if mother's education is imputed}
\item{weight}{NLS weight for 1976 cross-section}
\item{momdad14}{Indicator if lived with both parents at age 14}
\item{sinmom14}{Indicator if lived with single mother at age 14}
\item{step14}{Indicator if lived with step-parent at age 14}
\item{reg661}{Dummy for region 1 in 1966}
\item{reg662}{Dummy for region 2 in 1966}
\item{reg663}{Dummy for region 3 in 1966}
\item{reg664}{Dummy for region 4 in 1966}
\item{reg665}{Dummy for region 5 in 1966}
\item{reg666}{Dummy for region 6 in 1966}
\item{reg667}{Dummy for region 7 in 1966}
\item{reg668}{Dummy for region 8 in 1966}
\item{reg669}{Dummy for region 9 in 1966}
\item{south66}{Indicator if lived in the South in 1966}
\item{work76}{Indicator if worked in 1976}
\item{work78}{Indicator if worked in 1978}
\item{lwage76}{Log wage in 1976 (outliers trimmed)}
\item{lwage78}{Log wage in 1978 (outliers trimmed)}
\item{famed}{Combined education level of mother and father}
\item{black}{Indicator if Black}
\item{smsa76r}{Indicator if lived in SMSA in 1976}
\item{smsa78r}{Indicator if lived in SMSA in 1978}
\item{reg76r}{Indicator if lived in the South in 1976}
\item{reg78r}{Indicator if lived in the South in 1978}
\item{reg80r}{Indicator if lived in the South in 1980}
\item{smsa66r}{Indicator if lived in SMSA in 1966}
\item{wage76}{Raw wage in cents per hour in 1976}
\item{wage78}{Raw wage in cents per hour in 1978}
\item{wage80}{Raw wage in cents per hour in 1980}
\item{noint78}{Indicator if non-interviewed in 1978}
\item{noint80}{Indicator if non-interviewed in 1980}
\item{enroll76}{Indicator if enrolled in 1976}
\item{enroll78}{Indicator if enrolled in 1978}
\item{enroll80}{Indicator if enrolled in 1980}
\item{kww}{Knowledge of the World of Work (KWW) score}
\item{iq}{Normed IQ score}
\item{marsta76}{Married and with spouse present in 1976}
\item{marsta78}{Married and with spouse present in 1978}
\item{marsta80}{Married and with spouse present in 1980}
\item{libcrd14}{Indicator if had a library card at home at age 14}
}
}
\source{
Card, D. (1995). "Using Geographic Variation in College Proximity to Estimate the Return to Schooling." In L.N. Christofides, E.K. Grant, and R. Swidinsky (Eds.), \emph{Aspects of Labor Market Behaviour: Essays in Honour of John Vanderkamp}. Toronto: University of Toronto Press. Data available at \url{http://davidcard.berkeley.edu/data_sets.html}
}
\usage{
card1995
}
\description{
Education, earnings, family background, and regional variables from the 1976 cross-section of the National Longitudinal Survey (NLS) of young men.
}
\keyword{datasets}
