% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{chj2004}
\alias{chj2004}
\title{Cox, Hansen, and Jimenez (2004) Dataset on Private Transfers and Income}
\format{
A data frame with 8,685 observations and 28 variables.
\describe{
\item{tabroad}{Transfers received from abroad}
\item{tdomestic}{Transfers received domestically}
\item{tinkind}{Transfers received in-kind}
\item{tgifts}{Gifts and transfers out (gross)}
\item{income}{Total income}
\item{wage}{Wage and salary income}
\item{entrepreneourial}{Entrepreneurial income}
\item{rental}{Rental income}
\item{interest}{Interest income}
\item{pension}{Pension and retirement income}
\item{dividend}{Dividend income}
\item{primary}{Primary school graduate (dummy)}
\item{somesecondary}{Some secondary school (dummy)}
\item{secondary}{Secondary school graduate (dummy)}
\item{someuniversity}{Some university education (dummy)}
\item{university}{University graduate (dummy)}
\item{age}{Age of household head}
\item{female}{Female-headed household (dummy)}
\item{married}{Married household head (dummy)}
\item{child1}{Number of children aged 1 year or less}
\item{child7}{Number of children aged 1 to 7 years}
\item{child15}{Number of children aged 8 to 15 years}
\item{size}{Household size}
\item{bothwork}{Both husband and wife work (dummy)}
\item{notemployed}{Household head not employed (dummy)}
\item{marriedf}{Married female household head (dummy)}
\item{region}{Region code (0 through 13)}
\item{transfers}{Total transfers (tabroad + tdomestic + tinkind - tgifts)}
}
}
\source{
Cox, D., Hansen, B. E., & Jimenez, E. (2004). "How responsive are private transfers to income? Evidence from a laissez-faire economy." \emph{Journal of Public Economics}, 88(10), 2193-2219.
}
\usage{
chj2004
}
\description{
Private transfers, income sources, and household demographics for urban Filipino households. The dataset contains 8,685 observations, with the top 2\\% of income and observations with negative income removed.
}
\keyword{datasets}
