% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{ck1994}
\alias{ck1994}
\title{Card and Krueger (1994) Dataset on Minimum Wages and Employment}
\format{
A data frame with 820 observations and 26 variables, representing 410 stores surveyed twice.
\describe{
\item{store}{Unique store ID}
\item{chain}{Fast-food chain}
\item{co_owned}{Company-owned}
\item{state}{State}
\item{southj}{Located in southern New Jersey}
\item{centralj}{Located in central New Jersey}
\item{northj}{Located in northern New Jersey}
\item{pa1}{Located in the northeast suburbs of Philadelphia}
\item{pa2}{Located in the outside northeast suburbs of Philadelphia}
\item{shore}{Located on New Jersey shore}
\item{ncalls}{Number of call-backs (0 if contacted on first call)}
\item{empft}{Number of full-time employees}
\item{emppt}{Number of part-time employees}
\item{nmgrs}{Number of managers/assistant managers}
\item{wage_st}{Starting wage (dollars per hour)}
\item{inctime}{Months to usual first raise}
\item{firstinc}{Usual amount of first raise (dollars per hour)}
\item{meals}{Free/reduced price meal code}
\item{open}{Hour of opening}
\item{hoursopen}{Number of hours open per day}
\item{pricesoda}{Price of medium soda, including tax}
\item{pricefry}{Price of small fries, including tax}
\item{priceentree}{Price of entrée, including tax}
\item{nregisters}{Number of cash registers in store}
\item{nregisters11}{Number of registers open at 11am}
\item{time}{Survey wave indicator}
}
}
\source{
Card, D., & Krueger, A. B. (1994). "Minimum Wages and Employment: A Case Study of the Fast-Food Industry in New Jersey and Pennsylvania." \emph{American Economic Review}. Data available at \url{http://davidcard.berkeley.edu/data_sets.html}
}
\usage{
ck1994
}
\description{
Employment and wage characteristics in fast-food restaurants across New Jersey and Pennsylvania. The dataset includes information on store characteristics, wages, and employee counts, reorganized so that each row represents one observation per store across two survey waves.
}
\keyword{datasets}
