% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{cls}
\alias{cls}
\title{OLS estimation with HC2 cluster-robust standard errors}
\usage{
cls(y, x)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}
}
\description{
OLS estimation with HC2 cluster-robust standard errors
}
