% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{cmr2008}
\alias{cmr2008}
\title{Card, Mas, and Rothstein (2008) Segregation Dynamics Dataset}
\format{
A data frame with 48,643 observations (census tracts) and 46 variables.
\describe{
\item{geo2000}{Census tract code}
\item{unem_70}{Unemployment rate for population 16+ years, 1970}
\item{unem_80}{Unemployment rate for population 16+ years, 1980}
\item{unem_90}{Unemployment rate for population 16+ years, 1990}
\item{unem_00}{Unemployment rate for population 16+ years, 2000}
\item{msa}{Metropolitan Statistical Area (MSA) code}
\item{city}{Name of the Metropolitan Statistical Area (MSA)}
\item{pubtran_70}{Fraction of workers using public transportation, 1970}
\item{pubtran_80}{Fraction of workers using public transportation, 1980}
\item{pubtran_90}{Fraction of workers using public transportation, 1990}
\item{pop_70}{Population, 1970}
\item{pop_80}{Population, 1980}
\item{pop_90}{Population, 1990}
\item{pop_00}{Population, 2000}
\item{fr_white_70}{Fraction White, 1970}
\item{fr_white_80}{Fraction White, 1980}
\item{fr_white_90}{Fraction White, 1990}
\item{fr_white_00}{Fraction White, 2000}
\item{fr_min_70}{Fraction Minority, 1970}
\item{fr_min_80}{Fraction Minority, 1980}
\item{fr_min_90}{Fraction Minority, 1990}
\item{fr_min_00}{Fraction Minority, 2000}
\item{chg_white_7080}{Change in White share, 1970-1980}
\item{chg_white_8090}{Change in White share, 1980-1990}
\item{chg_white_9000}{Change in White share, 1990-2000}
\item{chg_min_7080}{Change in Minority share, 1970-1980}
\item{chg_min_8090}{Change in Minority share, 1980-1990}
\item{chg_min_9000}{Change in Minority share, 1990-2000}
\item{chg_pop_7080}{Change in population, 1970-1980}
\item{chg_pop_8090}{Change in population, 1980-1990}
\item{chg_pop_9000}{Change in population, 1990-2000}
\item{samp_70}{Indicator for inclusion in 1970 sample}
\item{samp_80}{Indicator for inclusion in 1980 sample}
\item{samp_90}{Indicator for inclusion in 1990 sample}
\item{faminc_70}{Log of mean family income, 1970}
\item{faminc_80}{Log of mean family income, 1980}
\item{faminc_90}{Log of mean family income, 1990}
\item{vac_70}{Fraction of vacant homes, 1970}
\item{vac_80}{Fraction of vacant homes, 1980}
\item{vac_90}{Fraction of vacant homes, 1990}
\item{rent_70}{Fraction of renter-occupied homes, 1970}
\item{rent_80}{Fraction of renter-occupied homes, 1980}
\item{rent_90}{Fraction of renter-occupied homes, 1990}
\item{oneunit_70}{Fraction of single-unit homes, 1970}
\item{oneunit_80}{Fraction of single-unit homes, 1980}
\item{oneunit_90}{Fraction of single-unit homes, 1990}
}
}
\source{
Card, D., Mas, A., & Rothstein, J. (2008). "Tipping and the Dynamics of Segregation." \emph{Quarterly Journal of Economics}. Data available at \url{https://berkeley.app.box.com/v/tippingreplication}
}
\usage{
cmr2008
}
\description{
Demographic, economic, and housing characteristics across U.S. census tracts from 1970 to 2000. The dataset includes variables related to population composition, income, housing, and transportation, reorganized from the original replication files.
}
\keyword{datasets}
