% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{cns}
\alias{cns}
\title{Constrained regression}
\usage{
cns(y, x, r)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}

\item{r}{constraint matrix}
}
\value{
matrix of constrained regression estimates
}
\description{
Constrained regression
}
