% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{ddk}
\alias{ddk}
\title{OLS estimation with DDK (2011) cluster-robust standard errors}
\usage{
ddk(y, x, z)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}

\item{z}{cluster variable}
}
\description{
OLS estimation with DDK (2011) cluster-robust standard errors
}
