% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{ds2004}
\alias{ds2004}
\title{DiTella and Schargrodsky (2004) Police Presence and Crime Dataset}
\format{
A data frame with 7,884 observations and 10 variables.
\describe{
\item{block}{Block ID}
\item{barrio}{Neighborhood}
\item{calle}{Street name}
\item{altura}{Street numbering}
\item{sameblock}{Indicator if there is a Jewish institution on the block}
\item{distance}{Number of blocks to the closest Jewish institution}
\item{public}{Indicator if there is a public building on the block}
\item{gasstation}{Indicator if there is a gas station on the block}
\item{bank}{Indicator if there is a bank on the block}
\item{thefts}{Total car thefts on the block during the month}
\item{month}{Month of observation (numeric)}
\item{oneblock}{Indicator if there is a Jewish institution one block away}
}
}
\source{
Di Tella, R., & Schargrodsky, E. (2004). "Do Police Reduce Crime?
Estimates Using the Allocation of Police Forces After a Terrorist Attack."
\emph{American Economic Review}. Data available at
\url{https://www.aeaweb.org/aer/data/march2004_ditella_data.zip}
}
\usage{
ds2004
}
\description{
Police presence, neighborhood characteristics, and crime levels across city
blocks in Buenos Aires. The data spans from April 1994 through December 1994,
with monthly observations on 7,884 city blocks.
}
\keyword{datasets}
