% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{emd}
\alias{emd}
\title{Efficient Minimum Distance (EMD) regression}
\usage{
emd(y, x, r)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}

\item{r}{constraint matrix}
}
\value{
matrix of EMD estimates
}
\description{
Efficient Minimum Distance (EMD) regression
}
