% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{fred_qd}
\alias{fred_qd}
\title{McCracken and Ng (2015) FRED-QD Quarterly Database for Macroeconomic Research}
\format{
A data frame with 236 observations and 249 variables:
\describe{
\item{gdpc1}{Real Gross Domestic Product}
\item{pcecc96}{Real Personal Consumption Expenditures}
\item{pcdgx}{Real Personal Consumption Expenditures: Durable Goods}
\item{pcesvx}{Real Personal Consumption Expenditures: Services excluding
food and energy}
\item{pcndx}{Real Personal Consumption Expenditures: Nondurable Goods}
\item{gpdic1}{Gross Private Domestic Investment}
\item{fpix}{Fixed Private Investment}
\item{y033rc1q027sbeax}{Real Gross Private Domestic Investment:
Residential}
\item{pnfix}{Private Nonresidential Fixed Investment}
\item{prfix}{Private Residential Fixed Investment}
\item{a014re1q156nbea}{Government Consumption Expenditures}
\item{gcec1}{Real Government Consumption Expenditures and Gross Investment}
\item{a823rl1q225sbea}{Real Government Spending}
\item{fgrecptx}{Federal Government Current Receipts}
\item{slcex}{State and Local Government Current Expenditures}
\item{expgsc1}{Exports of Goods and Services}
\item{impgsc1}{Imports of Goods and Services}
\item{dpic96}{Disposable Personal Income}
\item{outnfb}{Nonfarm Business Sector: Output}
\item{outbs}{Business Sector Output}
\item{outms}{Manufacturing Sector Output}
\item{indpro}{Industrial Production Index}
\item{ipfinal}{Industrial Production: Final Products}
\item{ipcongd}{Industrial Production: Consumer Goods}
\item{ipmat}{Industrial Production: Materials}
\item{ipdmat}{Industrial Production: Durable Materials}
\item{ipnmat}{Industrial Production: Nondurable Materials}
\item{ipdcongd}{Industrial Production: Durable Consumer Goods}
\item{ipb51110sq}{Industrial Production: Utilities}
\item{ipncongd}{Industrial Production: Nondurable Consumer Goods}
\item{ipbuseq}{Industrial Production: Business Equipment}
\item{ipb51220sq}{Industrial Production: Residential Utilities}
\item{tcu}{Total Capacity Utilization}
\item{cumfns}{Nonfarm Business Sector: Total Compensation}
\item{payems}{Total Nonfarm Payroll Employment}
\item{uspriv}{All Employees: Private Industries}
\item{manemp}{All Employees: Manufacturing}
\item{srvprd}{All Employees: Service-Providing Industries}
\item{usgood}{All Employees: Goods-Producing Industries}
\item{dmanemp}{All Employees: Durable Goods}
\item{ndmanemp}{All Employees: Nondurable Goods}
\item{uscons}{All Employees: Construction}
\item{usehs}{All Employees: Education and Health Services}
\item{usfire}{All Employees: Financial Activities}
\item{usinfo}{All Employees: Information}
\item{uspbs}{All Employees: Professional and Business Services}
\item{uslah}{All Employees: Leisure and Hospitality}
\item{usserv}{All Employees: Other Services}
\item{usmine}{All Employees: Mining and Logging}
\item{ustpu}{All Employees: Trade, Transportation, and Utilities}
\item{usgovt}{All Employees: Government}
\item{ustrade}{All Employees: Retail Trade}
\item{uswtrade}{All Employees: Wholesale Trade}
\item{ces9091000001}{All Employees: State Government Education}
\item{ces9092000001}{All Employees: State Government Excluding Education}
\item{ces9093000001}{All Employees: Local Government Education}
\item{ce16ov}{Civilian Employment: 16 Years and Over}
\item{civpart}{Civilian Labor Force Participation Rate}
\item{unrate}{Civilian Unemployment Rate}
\item{unratestx}{Unemployment Rate for State}
\item{unrateltx}{Unemployment Rate for Less than High School}
\item{lns14000012}{Unemployment Rate: Hispanic or Latino}
\item{lns14000025}{Unemployment Rate: 25-54 Years}
\item{lns14000026}{Unemployment Rate: 55 Years and Over}
\item{uemplt5}{Civilians Unemployed for Less Than 5 Weeks}
\item{uemp5to14}{Civilians Unemployed for 5 to 14 Weeks}
\item{uemp15t26}{Civilians Unemployed for 15 to 26 Weeks}
\item{uemp27ov}{Civilians Unemployed for 27 Weeks and Over}
\item{lns13023621}{Unemployment Rate for Bachelor’s Degree or Higher}
\item{lns13023557}{Unemployment Rate for Associate Degree Holders}
\item{lns13023705}{Unemployment Rate for Master’s Degree Holders}
\item{lns13023569}{Unemployment Rate for High School Graduates}
\item{lns12032194}{Labor Force Participation Rate for Bachelor’s Degree}
\item{hoabs}{Housing Starts: Buildings with 5 or More Units}
\item{hoams}{Housing Starts: All Buildings}
\item{hoanbs}{Nonfarm Business Sector: Hours Worked}
\item{awhman}{Average Weekly Hours: Manufacturing}
\item{awhnonag}{Average Weekly Hours: Nonagricultural Industries}
\item{awotman}{Average Weekly Overtime Hours: Manufacturing}
\item{hwix}{Help-Wanted Index}
\item{houst}{Housing Starts: Total}
\item{houst5f}{Housing Starts: 5-Unit Structures or More}
\item{permit}{New Private Housing Permits Authorized}
\item{houstmw}{Housing Starts: Midwest Region}
\item{houstne}{Housing Starts: Northeast Region}
\item{housts}{Housing Starts: South Region}
\item{houstw}{Housing Starts: West Region}
\item{cmrmtsplx}{Manufacturing and Trade Sales}
\item{rsafsx}{Residential Fixed Assets: Structures}
\item{amdmnox}{New Orders for Durable Goods}
\item{acognox}{New Orders: Consumer Goods and Materials}
\item{amdmuox}{Unfilled Orders: Durable Goods Industries}
\item{andenox}{Nondefense Capital Goods Excluding Aircraft}
\item{invcqrmtspl}{Inventories: Manufacturing and Trade Industries}
\item{pcectpi}{Personal Consumption Expenditures Chain-Type Price Index}
\item{pcepilfe}{PCE: Food and Energy}
\item{gdpctpi}{Gross Domestic Product Chain-Type Price Index}
\item{gpdictpi}{Gross Private Domestic Investment Chain-Type Price Index}
\item{ipdbs}{Industrial Production: Business Supplies}
\item{dgdsrg3q086sbea}{Durable Goods Expenditures}
\item{ddurrg3q086sbea}{Durable Goods Consumption Expenditures}
\item{dserrg3q086sbea}{Services Expenditures}
\item{dndgrg3q086sbea}{Nondurable Goods Expenditures}
\item{dhcerg3q086sbea}{Health Care Expenditures}
\item{dmotrg3q086sbea}{Motor Vehicle Expenditures}
\item{dfdhrg3q086sbea}{Food at Home Expenditures}
\item{dreqrg3q086sbea}{Recreational Goods Expenditures}
\item{dodgrg3q086sbea}{Other Durable Goods Expenditures}
\item{dfxarg3q086sbea}{Furniture and Fixtures Expenditures}
\item{dclorg3q086sbea}{Clothing and Footwear Expenditures}
\item{dgoerg3q086sbea}{Goods Expenditures}
\item{dongrg3q086sbea}{Nondurable Goods Consumption Expenditures}
\item{dhutrg3q086sbea}{Household Utilities Expenditures}
\item{dhlcrg3q086sbea}{Household Cleaning Expenditures}
\item{dtrsrg3q086sbea}{Transportation Services Expenditures}
\item{drcarg3q086sbea}{Recreational Services Expenditures}
\item{dfsarg3q086sbea}{Food Services and Accommodations}
\item{difsrg3q086sbea}{Insurance Services Expenditures}
\item{dotsrg3q086sbea}{Other Services Expenditures}
\item{cpiaucsl}{Consumer Price Index for All Urban Consumers: All Items}
\item{cpilfesl}{Consumer Price Index for All Urban Consumers: All Items
Less Food and Energy}
\item{wpsfd49207}{Producer Price Index by Commodity for Crude Materials}
\item{ppiaco}{Producer Price Index by Commodity}
\item{wpsfd49502}{Producer Price Index by Commodity for Intermediate
Materials}
\item{wpsfd4111}{Producer Price Index by Commodity for Finished Goods}
\item{ppiidc}{Producer Price Index: Industrial Commodities}
\item{wpsid61}{Producer Price Index for Finished Goods}
\item{wpu0531}{Producer Price Index for Gasoline}
\item{wpu0561}{Producer Price Index for Fuel Oil}
\item{oilpricex}{Crude Oil Prices}
\item{ahetpix}{Average Hourly Earnings of Production Employees}
\item{ces2000000008x}{Average Hourly Earnings: Construction}
\item{ces3000000008x}{Average Hourly Earnings: Manufacturing}
\item{comprms}{Compensation per Hour: Manufacturing}
\item{comprnfb}{Compensation per Hour: Nonfarm Business}
\item{rcphbs}{Real Compensation Per Hour}
\item{ophmfg}{Output per Hour: Manufacturing}
\item{ophnfb}{Output per Hour: Nonfarm Business}
\item{ophpbs}{Output per Hour: Private Business}
\item{ulcbs}{Unit Labor Cost: Business Sector}
\item{ulcmfg}{Unit Labor Cost: Manufacturing}
\item{ulcnfb}{Unit Labor Cost: Nonfarm Business}
\item{unlpnbs}{Unit Nonlabor Payments}
\item{fedfunds}{Effective Federal Funds Rate}
\item{tb3ms}{3-Month Treasury Bill Secondary Market Rate}
\item{tb6ms}{6-Month Treasury Bill Secondary Market Rate}
\item{gs1}{1-Year Treasury Constant Maturity Rate}
\item{gs10}{10-Year Treasury Constant Maturity Rate}
\item{mortgage30us}{30-Year Fixed Rate Mortgage Average}
\item{aaa}{Moody's AAA Corporate Bond Yield}
\item{baa}{Moody's BAA Corporate Bond Yield}
\item{baa10ym}{Baa Rated Corporate Bond Yield Spread Over 10-Year Treasury}
\item{mortg10yrx}{10-Year Treasury Note Average Yield}
\item{tb6m3mx}{6-Month Treasury Minus 3-Month Treasury Spread}
\item{gs1tb3mx}{1-Year Treasury Minus 3-Month Treasury Spread}
\item{gs10tb3mx}{10-Year Treasury Minus 3-Month Treasury Spread}
\item{cpf3mtb3mx}{Commercial Paper Minus 3-Month Treasury Spread}
\item{ambslrealx}{Adjusted Monetary Base}
\item{imfslx}{Monetary Base, Not Seasonally Adjusted}
\item{m1realx}{Real M1 Money Stock}
\item{m2realx}{Real M2 Money Stock}
\item{mzmrealx}{MZM Money Stock}
\item{busloansx}{Commercial and Industrial Loans at All Commercial Banks}
\item{consumerx}{Consumer Credit Outstanding}
\item{nonrevslx}{Nonrevolving Consumer Credit}
\item{reallnx}{Real Estate Loans}
\item{revolslx}{Revolving Consumer Credit}
\item{totalslx}{Total Loans and Leases}
\item{driwcil}{Deposits and Reserves in Financial Institutions}
\item{tabshnox}{Total Assets at Banks}
\item{tlbshnox}{Total Liabilities at Banks}
\item{liabpix}{Liabilities}
\item{tnwbshnox}{Total Net Worth of Households and Nonprofits}
\item{nwpix}{Net Worth at Market Value}
\item{taresax}{Treasury Assets and Reserves}
\item{hnoremq027sx}{Household Net Worth}
\item{tfaabshnox}{Total Financial Assets}
\item{vxoclsx}{CBOE S&P 100 Volatility Index}
\item{ussthpi}{All-Transactions House Price Index for the United States}
\item{spcs10rsa}{S&P/Case-Shiller 10-City Composite Home Price Index}
\item{spcs20rsa}{S&P/Case-Shiller 20-City Composite Home Price Index}
\item{twexmmth}{Trade Weighted U.S. Dollar Index: Major Currencies}
\item{exuseu}{Exchange Rate: U.S. Dollar to Euro}
\item{exszusx}{Switzerland / U.S. Foreign Exchange Rate}
\item{exjpusx}{Japan / U.S. Foreign Exchange Rate}
\item{exusukx}{U.K. / U.S. Foreign Exchange Rate}
\item{excausx}{Canada / U.S. Foreign Exchange Rate}
\item{umcsentx}{University of Michigan: Consumer Sentiment Index}
\item{usepuindxm}{U.S. Economic Policy Uncertainty Index}
\item{b020re1q156nbea}{Government Current Expenditures}
\item{b021re1q156nbea}{Net Government Saving}
\item{gfdegdq188s}{Federal Debt: Total Public Debt}
\item{gfdebtnx}{Gross Federal Debt as Percent of GDP}
\item{ipmansics}{Industrial Production: Manufacturing}
\item{ipb51222s}{Industrial Production: Fuels}
\item{ipfuels}{Industrial Production: Fuels}
\item{uempmean}{Average Duration of Unemployment}
\item{ces0600000007}{Average Weekly Hours: Goods-Producing Industries}
\item{totresns}{Total Reserves of Depository Institutions}
\item{nonborres}{Nonborrowed Reserves of Depository Institutions}
\item{gs5}{5-Year Treasury Constant Maturity Rate}
\item{tb3smffm}{3-Month Treasury Minus Fed Funds Rate Spread}
\item{t5yffm}{5-Year Treasury Minus Fed Funds Rate Spread}
\item{aaaffm}{Moody's AAA Corporate Bond Minus Fed Funds Rate Spread}
\item{wpsid62}{Producer Price Index by Commodity for Finished Goods:
Consumer}
\item{ppicmm}{Producer Price Index: Commodities}
\item{cpiappsl}{Consumer Price Index for All Urban Consumers}
\item{cpitrnsl}{CPI: Transportation}
\item{cpimedsl}{CPI: Medical Care}
\item{cusr0000sac}{CPI for All Urban Consumers: Commodities}
\item{cusr0000sad}{CPI for All Urban Consumers: Durables}
\item{cusr0000sas}{CPI for All Urban Consumers: Services}
\item{cpiulfsl}{CPI: All Items Less Food and Energy}
\item{cusr0000sa0l2}{CPI: Energy}
\item{cusr0000sa0l5}{CPI: Food}
\item{ces0600000008}{Average Hourly Earnings: Goods-Producing Industries}
\item{dtcolnvhfnm}{Consumer Motor Vehicle Loans Outstanding}
\item{dtcthfnm}{Total Consumer Loans Outstanding}
\item{invest}{Investment in Bank Credit}
\item{hwiuratiox}{Help-Wanted Index to Unemployed Ratio}
\item{claimsx}{Initial Claims for Unemployment Insurance}
\item{businvx}{Total Business Inventories}
\item{isratiox}{Inventory-to-Sales Ratio}
\item{conspix}{Consumer Credit to Personal Income Ratio}
\item{cp3m}{3-Month Commercial Paper Rate}
\item{compapff}{Commercial Paper Minus Fed Funds Rate Spread}
\item{permitne}{New Private Housing Permits Authorized: Northeast}
\item{permitmw}{New Private Housing Permits Authorized: Midwest}
\item{permits}{New Private Housing Permits Authorized}
\item{permitw}{New Private Housing Permits Authorized: West}
\item{nikkei225}{Nikkei 225 Stock Index}
\item{nasdaqcom}{NASDAQ Composite Index}
\item{cusr0000sehc}{CPI for All Urban Consumers: Education}
\item{tlbsnncbx}{Total Liabilities for Nonfinancial Corporate Business}
\item{tlbsnncbbdix}{Total Liabilities for Nonfinancial Corporate Business,
Billion Dollars}
\item{ttaabsnncbx}{Total Assets for Nonfinancial Corporate Business}
\item{tnwmvbsnncbx}{Total Net Worth for Nonfinancial Corporate Business}
\item{tnwmvbsnncbbdix}{Total Net Worth for Nonfinancial Corporate Business,
Billion Dollars}
\item{tlbsnnbx}{Total Liabilities for Nonfinancial Noncorporate Business}
\item{tlbsnnbbdix}{Total Liabilities for Nonfinancial Noncorporate
Business, Billion Dollars}
\item{tabsnnbx}{Total Assets for Nonfinancial Noncorporate Business}
\item{tnwbsnnbx}{Total Net Worth for Nonfinancial Noncorporate Business}
\item{tnwbsnnbbdix}{Total Net Worth for Nonfinancial Noncorporate Business,
Billion Dollars}
\item{cncfx}{Real Personal Consumption Expenditures: Non-Durables and
Durables}
\item{sp500}{S&P 500 Index}
\item{spindust}{S&P Industrial Index}
\item{spdivyield}{S&P Dividend Yield}
\item{spperatio}{S&P Price-Earnings Ratio}
\item{time}{Time Period}
}
}
\source{
McCracken, M. W., & Ng, S. "FRED-QD: A Quarterly Database for
Macroeconomic Research." Data available at
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
fred_qd
}
\description{
Quarterly collection of major U.S. macroeconomic variables maintained by the
St. Louis Federal Reserve Bank. This dataset provides key economic indicators
across sectors such as output, labor, housing, consumption, interest rates,
money, productivity, prices, and exchange rates. Includes data from 1959
onward.
}
\details{
Most of the variables are provided in the most basic format, so
transformations are required for typical uses (e.g., for growth rates).
}
\keyword{datasets}
