% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{influence}
\alias{influence}
\title{Influence function}
\usage{
influence(y, x)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}
}
\value{
numerical value of influence function
}
\description{
Influence function
}
