% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{invest1993}
\alias{invest1993}
\title{Hall and Hall (1993) Dataset Documentation}
\format{
A data frame with 27,566 observations and 23 variables:
\describe{
\item{cusip}{Committee on Uniform Security Identification Procedures (CUSIP) code, identifying each firm. The first six digits represent the firm code number.}
\item{year}{The year of the data, represented as a two-digit value.}
\item{pstar}{The present discounted value (PDV) of dividends anticipated on a firm's common stock in the future, discounted at the risk-free rate plus an overall risk premium (approximately 4 percent).}
\item{z0}{The first term in the linearized expression for z (as outlined in Appendix A of the paper).}
\item{pricef}{The actual price of common stock at the end of the fiscal year.}
\item{divf}{The dividends paid during the fiscal year.}
\item{rnda}{Research and development (R&D) expenditure to assets ratio.}
\item{adva}{Advertising expenditure to assets ratio.}
\item{fyr}{The month in which the fiscal year ends, ranging from 1 (January) to 12 (December).}
\item{ardsic}{A modified 2.5-digit Standard Industrial Classification (SIC) code, roughly similar to those used in Bound, Cummins, Griliches, Hall, and Jaffe (1984) in "Who Does R&D and Who Patents?"}
\item{exityr}{The year the firm exited from the sample.}
\item{inva}{Investment to assets ratio, reflecting the proportion of investment relative to the company's total assets.}
\item{cfa}{Cash flow to assets ratio, indicating the cash-generating efficiency of the assets.}
\item{debta}{Long-term debt to assets ratio, a measure of a company's financial leverage.}
\item{sales}{Annual sales revenue (in millions of dollars).}
\item{netcap}{The adjusted book value of assets, accounting for plant, property, equipment, inventories, and other components, with inflation adjustments as per Brainard, Shoven, and Weiss's approach.}
\item{earnsh}{Reported earnings per share (EPS).}
\item{nyseamex}{Indicator variable set to 1 if the firm is listed on the New York Stock Exchange (NYSE) or American Stock Exchange (AMEX), otherwise 0.}
\item{h0}{Instrumental variable as specified in equation B-5 of the referenced paper.}
\item{h1}{Another instrumental variable as specified in equation B-5 of the referenced paper.}
\item{vala}{Total market value to assets ratio, often referred to as Tobin's Q.}
\item{oneper}{One-period rate of return, adjusted for non-diversifiable discounting.}
\item{sharef}{Shares of common stock outstanding, represented in thousands.}
}
}
\source{
Bronwyn H. Hall and Robert E. Hall, "The Value and Performance of U.S. Corporations," Brookings Papers on Economic Activity, 1993 (1), 1-50. Data sourced from \url{https://eml.berkeley.edu/~bhhall/bhdata.html}.
}
\usage{
invest1993
}
\description{
Detailed financial metrics for U.S. corporations, focusing on investment, financial ratios, and stock performance. Includes essential economic indicators that capture corporate financial health and performance for each year. Data spans multiple years and includes various financial metrics essential for corporate analysis.
}
\details{
All values are nominal and in millions of dollars unless stated otherwise. Stock values are end-of-year.
}
\keyword{datasets}
