% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{nerlove1963}
\alias{nerlove1963}
\title{Nerlove (1963) Electricity Supply Dataset}
\format{
A data frame with 145 observations on the following variables:
\describe{
\item{cost}{Total Cost in million dollars}
\item{output}{Electricity output in billion kilowatt-hours (kWh)}
\item{plabor}{Unit price of labor, represented as wage rate in dollars per
hour}
\item{pcapital}{Unit price of capital, given as an index}
\item{pfuel}{Unit price of fuel in cents per million BTU (British Thermal
Units)}
}
}
\source{
Nerlove, Marc. "Returns to Scale in Electricity Supply" (1963),
\emph{Measurement in Economics}, C. Christ et al., eds.
}
\usage{
nerlove1963
}
\description{
Costs and production for 145 American electric companies in 1955.
}
\keyword{datasets}
