% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{ols}
\alias{ols}
\title{OLS estimation}
\usage{
ols(y, x)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}
}
\value{
matrix of OLS estimates
}
\description{
OLS estimation
}
