% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{ols2}
\alias{ols2}
\title{OLS estimation with standard errors (Homoskedastic, White, HC1, HC2, HC3)}
\usage{
ols2(y, x)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}
}
\value{
matrix of OLS estimates with standard errors
}
\description{
OLS estimation with standard errors (Homoskedastic, White, HC1, HC2, HC3)
}
