% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\name{ols3}
\alias{ols3}
\title{OLS estimation with standard errors}
\usage{
ols3(y, x)
}
\arguments{
\item{y}{dependent variable}

\item{x}{independent variables}
}
\value{
matrix of OLS estimates with standard errors
}
\description{
OLS estimation with standard errors
}
