% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen-package.R
\docType{data}
\name{rr2010}
\alias{rr2010}
\title{Reinhart and Rogoff (2010) Growth in a Time of Debt Dataset}
\format{
A data frame with observations on the following variables:
\describe{
\item{year}{Calendar year, ranging from 1791 to 2009}
\item{debt}{Debt-to-GDP ratio, expressed as a percentage}
\item{gdp}{Annual GDP growth rate, expressed as a percentage}
\item{inflation}{Annual inflation rate, expressed as a percentage}
}
}
\source{
Reinhart, C. M., & Rogoff, K. S. (2010). "Growth in a Time of Debt." \emph{American Economic Review: Papers and Proceedings}, 573-578. Data available at \href{https://scholar.harvard.edu/files/rogoff/files/41_data.xlsx}{Ken Rogoff’s website}.
}
\usage{
rr2010
}
\description{
United States' economic indicators, with a focus on the relationship between debt levels and economic growth. The data cover a broad historical period from 1791 to 2009, extracted from the source files on Ken Rogoff's website.
}
\keyword{datasets}
